refactor: Rename main module to cli.py and update downstream usages
