+++ tests/test_cli.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 import pytest
 
-from pystormtracker.stormtracker import main, parse_args, run_tracker
+from pystormtracker.cli import main, parse_args, run_tracker
 
 from pystormtracker.utils.data_utils import fetch_era5_msl
