+++ tests/test_integration.py
@@ -8,8 +8,9 @@ from unittest.mock import patch
 
 import pytest
 
+from pystormtracker.io.imilast import read_imilast
 from pystormtracker.models.tracks import Tracks
-from pystormtracker.stormtracker import main
+from pystormtracker.cli import main
 
 from pystormtracker.utils.data_utils import fetch_era5_msl, fetch_era5_vo850
 
@@ -17,7 +18,7 @@ from pystormtracker.utils.data_utils import fetch_era5_msl, fetch_era5_vo850
 def run_command_direct(cmd_args: list[str], use_mpi: bool = False) -> None:
     """Utility to run the tracker directly via function calls or MPI subprocess."""
     if use_mpi:
-        base_cmd = f"{sys.executable} -m pystormtracker.stormtracker"
+        base_cmd = f"{sys.executable} -m pystormtracker.cli"
         full_cmd = f"mpiexec -n 2 {base_cmd} {' '.join(cmd_args)}"
         subprocess.run(full_cmd, shell=True, check=True, capture_output=True)
         return
@@ -47,8 +48,8 @@ def compare_tracks(
     intensity_tol: float = 1e-4,
 ) -> None:
     """Compares two tracking files for equality using the Tracks class."""
-    t1 = Tracks.from_imilast(file1)
-    t2 = Tracks.from_imilast(file2)
+    t1 = read_imilast(file1)
+    t2 = read_imilast(file2)
     t1.compare(
         t2,
         length_diff_tol=length_diff_tol,
