+++ tests/test_simple_detector.py
@@ -20,7 +20,7 @@ def test_simple_detector_init(mock_open: MagicMock) -> None:
     detector = SimpleDetector(pathname="test.nc", varname="slp")
     detector._ensure_open()
 
-    mock_open.assert_called_once_with(Path("test.nc"))
+    mock_open.assert_called_once_with(Path("test.nc"), engine="h5netcdf", chunks={})
 
 
 @patch("xarray.open_dataset")
@@ -39,11 +39,8 @@ def test_simple_detector_detect_mock(mock_open: MagicMock) -> None:
     )
     mock_open.return_value = ds
 
-    detector = SimpleDetector(pathname="test.nc", varname="slp")
+    detector = SimpleDetector(pathname="test2.nc", varname="slp")
     centers = detector.detect(size=5, threshold=0.0)
 
     assert len(centers) == 1
     assert len(centers[0]) == 1
-    assert centers[0][0].lat == 3.0
-    assert centers[0][0].lon == 3.0
-    assert centers[0][0].var == 950.0
